Configure GitLab CI to automate week 3 testing
Multiple CI jobs are added into the `.gitlab-ci.yml` file so that
the works related to development of week 3 tutorial by TA and
the exercise worked by student can be automatically tested.

Flake8 is also added into the required Python dependencies. It is a
code style enforcement tool (linter). It will be used to ensure that
all Python code comply to PEP 8 standard.